feat: add condition checking when publish

+++ apps-api/backoffice/src/modules/admin/election/services.ts
@@ -47,6 +47,80 @@ export class AdminElectionService {
     private readonly ballotCryptoService: BallotCryptoService
   ) {}
 
+  private checkIsAllowedToPublish(election: Election) {
+    const now = new Date()
+    if (election.publishDate) {
+      return err({
+        code: InternalErrorCode.ELECTION_ALREADY_PUBLISH,
+        message: `Election is already published`,
+      })
+    }
+
+    if (election.type !== ElectionType.ONLINE) {
+      if (!election.location || !election.locationMapUrl) {
+        return err({
+          code: InternalErrorCode.ELECTION_NOT_ALLOWED_TO_PUBLISH,
+          message: `Election of type ${election.type} must have location info to allow publishing`,
+        })
+      }
+    }
+
+    if (election.type !== ElectionType.ONSITE) {
+      if (election.keysStatus !== ElectionKeysStatus.CREATED) {
+        return err({
+          code: InternalErrorCode.ELECTION_NOT_ALLOWED_TO_PUBLISH,
+          message: `Election of type ${election.type} must have created keys to allow publishing`,
+        })
+      }
+    }
+
+    if (election.type === ElectionType.HYBRID) {
+      if (!election.openRegister || !election.closeRegister) {
+        return err({
+          code: InternalErrorCode.ELECTION_NOT_ALLOWED_TO_PUBLISH,
+          message: `Election of type ${election.type} must have registration period to allow publishing`,
+        })
+      }
+
+      if (election.openRegister <= now) {
+        return err({
+          code: InternalErrorCode.ELECTION_NOT_ALLOWED_TO_PUBLISH,
+          message: `Election registration period should be in the future to allow publishing`,
+        })
+      }
+
+      if (election.closeRegister <= election.openRegister) {
+        return err({
+          code: InternalErrorCode.ELECTION_NOT_ALLOWED_TO_PUBLISH,
+          message: `Election close registration must be after open registration`,
+        })
+      }
+
+      if (election.openVoting <= election.closeRegister) {
+        return err({
+          code: InternalErrorCode.ELECTION_NOT_ALLOWED_TO_PUBLISH,
+          message: `Election open voting must be after close registration`,
+        })
+      }
+    }
+
+    if (election.openVoting <= now) {
+      return err({
+        code: InternalErrorCode.ELECTION_NOT_ALLOWED_TO_PUBLISH,
+        message: `Election voting period should be in the future to allow publishing`,
+      })
+    }
+
+    if (election.closeVoting <= election.openVoting) {
+      return err({
+        code: InternalErrorCode.ELECTION_NOT_ALLOWED_TO_PUBLISH,
+        message: `Election close voting must be after open voting`,
+      })
+    }
+
+    return ok()
+  }
+
   private checkIsDraftElection(election: Election) {
     if (election.publishDate) {
       return err({
@@ -58,6 +132,8 @@ export class AdminElectionService {
   }
 
   private checkIsElectionAllowedToModified(election: Election, now: Date) {
+    if (election.publishDate === null) return ok()
+
     const isOpenVote = now >= election.openVoting
     if (isOpenVote) {
       return err({
@@ -294,20 +370,31 @@ export class AdminElectionService {
   }
 
   async cancelElection(electionId: string) {
-    const destroyKeysResult = await this.ballotCryptoService.destroyElectionKeys(electionId)
-    if (destroyKeysResult.isErr()) return err(destroyKeysResult.error)
+    const electionDetails = await this.adminElectionRepository.getElectionById(electionId)
+    if (electionDetails.isErr()) {
+      return mapRepositoryError(electionDetails.error, {
+        RECORD_NOT_FOUND: {
+          code: InternalErrorCode.ELECTION_NOT_FOUND,
+          message: `Cannot found election id: ${electionId}`,
+        },
+      })
+    }
+
+    const isDestroyKey = electionDetails.value.type !== ElectionType.ONSITE
+
+    if (isDestroyKey) {
+      const destroyKeysResult = await this.ballotCryptoService.destroyElectionKeys(electionId)
+      if (destroyKeysResult.isErr()) return err(destroyKeysResult.error)
+    }
 
     const cancelResult = await this.adminElectionRepository.cancelElectionById(electionId)
     if (cancelResult.isErr()) {
-      const restoreKeysResult = await this.ballotCryptoService.restoreKeys(electionId)
-      if (restoreKeysResult.isErr()) return err(restoreKeysResult.error)
+      if (isDestroyKey) {
+        const restoreKeysResult = await this.ballotCryptoService.restoreKeys(electionId)
+        if (restoreKeysResult.isErr()) return err(restoreKeysResult.error)
+      }
 
-      return mapRepositoryError(cancelResult.error, {
-        RECORD_NOT_FOUND: {
-          code: InternalErrorCode.ELECTION_NOT_FOUND,
-          message: `Cannot Found Election id: ${electionId}`,
-        },
-      })
+      return mapRepositoryError(cancelResult.error)
     }
 
     return ok()
@@ -324,14 +411,25 @@ export class AdminElectionService {
       })
     }
 
-    const checkResult = this.checkIsDraftElection(electionResult.value)
+    const checkResult = this.checkIsAllowedToPublish(electionResult.value)
     if (checkResult.isErr()) return err(checkResult.error)
+
+    const isDestroyKey = electionResult.value.type === ElectionType.ONSITE
+    if (isDestroyKey) {
+      const destroyKeysResult = await this.ballotCryptoService.destroyElectionKeys(electionId)
+      if (destroyKeysResult.isErr()) return err(destroyKeysResult.error)
+    }
 
     const publishResult = await this.adminElectionRepository.publishElectionById(
       electionId,
-      publishDate
+      publishDate,
+      isDestroyKey
     )
     if (publishResult.isErr()) {
+      if (isDestroyKey) {
+        const restoreResult = await this.ballotCryptoService.restoreKeys(electionId)
+        if (restoreResult.isErr()) return err(restoreResult.error)
+      }
       return mapRepositoryError(publishResult.error)
     }
